+++ packages/ui/src/hooks/useCoinConversion.ts
@@ -1,6 +1,7 @@
 import { useContext } from 'react'
 import useSWR from 'swr'
 import { CoinGecko, ProviderOptionsContext } from '../ReservoirKitProvider'
+import useCoinIds from './useCoinIds'
 
 const createBaseUrl = (config: CoinGecko | undefined) => {
   if (config?.proxy) return `${config.proxy}?`
@@ -17,22 +18,30 @@ export default function (
   id: string = ''
 ): { price: number; symbol: string }[] {
   const providerOptionsContext = useContext(ProviderOptionsContext)
-
+  const { data: coinIds } = useCoinIds()
   const baseUrl = createBaseUrl(providerOptionsContext?.coinGecko)
 
-  if (symbols.includes(',')) {
+  if (id.length === 0) {
     id = symbols
       .split(',')
-      .map((id) => providerOptionsContext.coinGecko?.coinIds?.[id])
+      .reduce((ids: string[], symbol: string) => {
+        const normalizedSymbol = symbol.toLowerCase()
+        if (providerOptionsContext.coinGecko?.coinIds?.[normalizedSymbol]) {
+          ids.push(providerOptionsContext.coinGecko.coinIds[normalizedSymbol])
+        } else {
+          const coins = coinIds[normalizedSymbol]
+          if (coins?.length === 1) {
+            ids.push(coins[0].id)
+          }
+        }
+
+        return ids
+      }, [])
       .join(',')
-  } else {
-    id = id ? id : providerOptionsContext.coinGecko?.coinIds?.[symbols] || ''
   }
 
   const { data } = useSWR(
-    vs_currency
-      ? `${baseUrl}vs_currency=${vs_currency}&symbols=${symbols}&ids=${id}`
-      : null,
+    vs_currency ? `${baseUrl}vs_currency=${vs_currency}&ids=${id}` : null,
     null,
     {
       refreshInterval: 60 * 1000 * 5, //5m Interval
@@ -40,11 +49,17 @@ export default function (
   )
 
   if (data && data.length > 0) {
-    return data.map((conversion: any) => ({
-      price: conversion.current_price,
-      symbol: (conversion.symbol || '').toUpperCase(),
-      coinGeckoId: conversion.id,
-    }))
+    return data
+      .filter((conversion: any) =>
+        symbols
+          .split(',')
+          .some((symbol) => symbol.toLowerCase().includes(conversion.symbol))
+      )
+      .map((conversion: any) => ({
+        price: conversion.current_price,
+        symbol: (conversion.symbol || '').toUpperCase(),
+        coinGeckoId: conversion.id,
+      }))
   }
   return []
 }